Remove GPT-2 references and use SIMONEModel
