--- SIM-ONE Training/sim_one_training/__init__.py
@@ -1,2 +1,0 @@
-"""SIM-ONE training package placeholder."""
-
