+++ docs/prioritary_mvlm.md
@@ -2,9 +2,9 @@
 
 ## Architecture
 
-`PrioritaryMVLM` wraps a minimal GPT-2 language model provided by the
-`transformers` library.  The model can be paired with `PrioritaryTokenizer`
-and trained using `PrioritaryTrainer` for experiments and fine-tuning.
+`PrioritaryMVLM` builds on the lightweight `SIMONEModel` transformer.
+The model pairs with `PrioritaryTokenizer` and can be trained using
+`PrioritaryTrainer` for experiments and fine-tuning.
 
 ## Training
 
@@ -14,13 +14,13 @@ For a complete walkthrough from environment setup to sampling, see the
 1. Prepare a dataset returning token ID tensors.
 2. Initialise the tokenizer and model:
    ```python
-   from prioritary_mvlm import PrioritaryTokenizer, PrioritaryMVLM
-   tokenizer = PrioritaryTokenizer.from_pretrained("gpt2")
-   model = PrioritaryMVLM()
+   from prioritary_mvlm import PrioritaryTokenizer, SIMONEModel
+   tokenizer = PrioritaryTokenizer()
+   model = SIMONEModel(vocab_size=len(tokenizer))
    ```
 3. Use the trainer to run a single training epoch:
    ```python
    from prioritary_mvlm import PrioritaryTrainer
-   trainer = PrioritaryTrainer(model, dataset)
-   trainer.train_epoch()
+   trainer = PrioritaryTrainer(data_dir="data", output_dir="checkpoints")
+   trainer.train()
    ```
