+++ prioritary_mvlm/__init__.py
@@ -1,13 +1,21 @@
-from .model import PrioritaryMVLM
+from simone_transformer.simone_model import SIMONEModel
 from .tokenizer import PrioritaryTokenizer
 from .trainer import PrioritaryTrainer
 from .config import PrioritaryConfig
 from .dataset import WeightedTextDataset
+from .losses import (
+    compute_policy_loss,
+    compute_memory_loss,
+    compute_energy_loss,
+)
 
 __all__ = [
-    "PrioritaryMVLM",
+    "SIMONEModel",
     "PrioritaryTokenizer",
     "PrioritaryTrainer",
     "PrioritaryConfig",
     "WeightedTextDataset",
+    "compute_policy_loss",
+    "compute_memory_loss",
+    "compute_energy_loss",
 ]
