+++ prioritary_mvlm/trainer.py
@@ -1,39 +1,27 @@
-"""Training utilities for the Prioritary MVLM.
-
-This module provides a lightweight trainer that mirrors the behaviour of
-``MVLMTrainer`` in the repository.  It supports multi‑epoch training with
-gradient accumulation, gradient clipping and an AdamW optimizer with a simple
-linear warmup schedule.  Training examples can carry metadata scores that
-indicate their priority; these scores are used to scale the loss and emphasise
-important samples.
-
-The trainer also includes helpers for evaluation, text generation and
-checkpoint management.  Logging is performed to the console and, optionally, a
-file so that training progress can be monitored easily.
-"""
+"""Training utilities for the SIM-ONE transformer."""
 
 from __future__ import annotations
 
 import logging
 import math
-from dataclasses import asdict
 from pathlib import Path
 from typing import Dict, Optional, Tuple
 
 import torch
+import torch.nn.functional as F
 from torch.optim import AdamW
 from torch.optim.lr_scheduler import LinearLR
 from torch.utils.data import DataLoader
-from transformers import GPT2Config
 
 from .config import PrioritaryConfig
 from .dataset import WeightedTextDataset
-from .model import PrioritaryMVLM
 from .tokenizer import PrioritaryTokenizer
+from simone_transformer.simone_model import SIMONEModel
+from .losses import compute_policy_loss, compute_memory_loss, compute_energy_loss
 
 
 class PrioritaryTrainer:
-    """Trainer for ``PrioritaryMVLM`` models."""
+    """Trainer for SIM-ONE models with priority-weighted loss."""
 
     def __init__(
         self,
@@ -42,28 +30,10 @@ class PrioritaryTrainer:
         config: Optional[PrioritaryConfig] = None,
         log_file: str | None = None,
     ) -> None:
-        """Initialise the trainer and all required components.
-
-        Parameters
-        ----------
-        data_dir:
-            Directory containing ``.txt`` and corresponding ``.json`` metadata
-            files.
-        output_dir:
-            Directory where checkpoints will be written.
-        config:
-            Optional training configuration.  When omitted a default
-            :class:`PrioritaryConfig` instance is created.
-        log_file:
-            Optional path to a log file for persisting metrics.
-        """
-
         self.config = config or PrioritaryConfig()
         self.output_dir = Path(output_dir)
         self.output_dir.mkdir(parents=True, exist_ok=True)
 
-        # ------------------------------------------------------------------
-        # Logging
         self.logger = logging.getLogger(self.__class__.__name__)
         self.logger.setLevel(logging.INFO)
         handlers = [logging.StreamHandler()]
@@ -74,26 +44,19 @@ class PrioritaryTrainer:
             h.setFormatter(fmt)
             self.logger.addHandler(h)
 
-        # ------------------------------------------------------------------
         # Tokenizer and model
-        self.tokenizer = PrioritaryTokenizer.from_pretrained("gpt2")
-        self.tokenizer.pad_token = self.tokenizer.eos_token
-
-        model_cfg = GPT2Config(
-            vocab_size=self.config.vocab_size,
-            n_layer=self.config.n_layer,
-            n_head=self.config.n_head,
-            n_embd=self.config.n_embd,
-            n_positions=self.config.max_length,
-            bos_token_id=50256,
-            eos_token_id=50256,
-            pad_token_id=50256,
+        self.tokenizer = PrioritaryTokenizer()
+        self.model = SIMONEModel(
+            vocab_size=len(self.tokenizer),
+            hidden_dim=self.config.hidden_dim,
+            num_heads=self.config.num_heads,
+            ff_dim=self.config.ff_dim,
+            num_layers=self.config.num_layers,
+            max_seq_len=self.config.max_length,
         )
-        self.model = PrioritaryMVLM(model_cfg)
         self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         self.model.to(self.device)
 
-        # ------------------------------------------------------------------
         # Dataset and dataloader
         self.dataset = WeightedTextDataset(data_dir, self.tokenizer, self.config)
         self.dataloader = DataLoader(
@@ -102,7 +65,6 @@ class PrioritaryTrainer:
             shuffle=True,
         )
 
-        # ------------------------------------------------------------------
         # Optimizer and scheduler
         self.optimizer = AdamW(
             self.model.parameters(),
@@ -120,28 +82,38 @@ class PrioritaryTrainer:
 
     # ------------------------------------------------------------------
     def compute_weighted_loss(self, batch: Dict) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Return the weighted and base loss for a batch.
-
-        ``WeightedTextDataset`` supplies a list of metadata dictionaries for
-        each batch under the ``"metadata"`` key.  Each dictionary may contain a
-        ``"priority"`` or ``"score"`` field.  Higher numbers indicate that the
-        sample should contribute more strongly to the training signal.
-        """
-
         input_ids = batch["input_ids"].to(self.device)
         labels = batch["labels"].to(self.device)
-        outputs = self.model(input_ids=input_ids, labels=labels)
-        base_loss = outputs.loss
 
-        priorities = [m.get("priority", m.get("score", 1.0)) for m in batch["metadata"]]
-        weight = sum(priorities) / len(priorities)
-        weighted_loss = base_loss * weight
-        return weighted_loss, base_loss
+        logits, aux = self.model(input_ids)
+        vocab_size = len(self.tokenizer)
+        mle_loss = F.cross_entropy(logits.view(-1, vocab_size), labels.view(-1))
+        policy_loss = compute_policy_loss(aux["policy_logits"][-1])
+        memory_loss = compute_memory_loss(aux["memory_signals"][-1])
+        energy_loss = compute_energy_loss(logits)
+
+        total = (
+            mle_loss
+            + self.config.lambda_policy * policy_loss
+            + self.config.lambda_memory * memory_loss
+            + self.config.lambda_energy * energy_loss
+        )
+
+
+        metadata = batch["metadata"]
+        if isinstance(metadata, dict):
+            if isinstance(metadata.get("priority"), torch.Tensor):
+                weight = metadata.get("priority").float().mean().item()
+            else:
+                weight = float(metadata.get("priority", metadata.get("score", 1.0)))
+        else:
+            priorities = [m.get("priority", m.get("score", 1.0)) for m in metadata]
+            weight = sum(priorities) / len(priorities)
+        weighted_loss = total * weight
+        return weighted_loss, mle_loss
 
     # ------------------------------------------------------------------
     def train(self, num_epochs: Optional[int] = None) -> None:
-        """Execute the training loop."""
-
         epochs = num_epochs or self.config.num_epochs
         for _ in range(epochs):
             self.logger.info(
@@ -167,7 +139,7 @@ class PrioritaryTrainer:
 
                     if self.global_step % self.config.log_interval == 0:
                         self.logger.info(
-                            "step=%d loss=%.4f base_loss=%.4f",  # pylint: disable=logging-format-interpolation
+                            "step=%d loss=%.4f mle=%.4f",
                             self.global_step,
                             loss.item() * self.config.gradient_accumulation_steps,
                             base_loss.item(),
@@ -191,8 +163,6 @@ class PrioritaryTrainer:
 
     # ------------------------------------------------------------------
     def evaluate(self, max_batches: int = 100) -> Tuple[float, float]:
-        """Evaluate the model on a subset of the training data."""
-
         self.model.eval()
         total_loss = 0.0
         batches = 0
@@ -209,50 +179,23 @@ class PrioritaryTrainer:
         return avg_loss, perplexity
 
     # ------------------------------------------------------------------
-    def generate_sample(
-        self, prompt: str = "In the beginning", max_length: int = 50
-    ) -> str:
-        """Generate a sample of text from the model."""
-
-        self.model.eval()
-        with torch.no_grad():
-            input_ids = self.tokenizer.encode(prompt, return_tensors="pt").to(self.device)
-            output = self.model.generate(
-                input_ids,
-                max_length=max_length,
-                num_return_sequences=1,
-                temperature=0.8,
-                do_sample=True,
-                pad_token_id=self.tokenizer.eos_token_id,
-            )
-        return self.tokenizer.decode(output[0], skip_special_tokens=True)
-
-    # ------------------------------------------------------------------
-    def save_checkpoint(self, name: str = "checkpoint.pt") -> Path:
-        """Persist training state to ``output_dir``."""
-
-        path = self.output_dir / name
-        checkpoint = {
-            "model_state": self.model.state_dict(),
-            "optimizer_state": self.optimizer.state_dict(),
-            "scheduler_state": self.scheduler.state_dict(),
-            "config": asdict(self.config),
+    def save_checkpoint(self, filename: str) -> None:
+        path = self.output_dir / filename
+        torch.save({
+            "model": self.model.state_dict(),
+            "optimizer": self.optimizer.state_dict(),
+            "scheduler": self.scheduler.state_dict(),
+            "config": self.config,
+            "step": self.global_step,
             "epoch": self.epoch,
-            "global_step": self.global_step,
-        }
-        torch.save(checkpoint, path)
-        self.logger.info("Saved checkpoint to %s", path)
-        return path
+        }, path)
 
     # ------------------------------------------------------------------
-    def load_checkpoint(self, path: str | Path) -> None:
-        """Load a previously saved checkpoint."""
-
-        checkpoint = torch.load(path, map_location=self.device)
-        self.model.load_state_dict(checkpoint["model_state"])
-        self.optimizer.load_state_dict(checkpoint["optimizer_state"])
-        self.scheduler.load_state_dict(checkpoint["scheduler_state"])
-        self.epoch = checkpoint.get("epoch", 0)
-        self.global_step = checkpoint.get("global_step", 0)
-        self.logger.info("Loaded checkpoint from %s", path)
-
+    def load_checkpoint(self, filename: str) -> None:
+        path = self.output_dir / filename
+        data = torch.load(path, map_location=self.device)
+        self.model.load_state_dict(data["model"])
+        self.optimizer.load_state_dict(data["optimizer"])
+        self.scheduler.load_state_dict(data["scheduler"])
+        self.global_step = data["step"]
+        self.epoch = data["epoch"]
